pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Toktamysov Timur",email = "timur.toktamysov@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.15.2,<2.0.0)",
    "annotated-types (>=0.7.0,<0.8.0)",
    "anyio (>=4.9.0,<5.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "click (>=8.1.8,<9.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "greenlet (>=3.2.0,<4.0.0)",
    "h11 (>=0.14.0,<0.15.0)",
    "idna (>=3.10,<4.0)",
    "mako (>=1.3.10,<2.0.0)",
    "markupsafe (>=3.0.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sniffio (>=1.3.1,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "starlette (>=0.46.2,<0.47.0)",
    "typing-inspection (>=0.4.0,<0.5.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "uvicorn (>=0.34.1,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

